Agent/DeepResearch/requirement.txt

python-dotenv
streamlit
pydantic_settings
langdetect
tavily-python








